chore: architecture access cleanup
