+++ contexts/index.ts
@@ -1,3 +1,5 @@
+export { BalancesProvider, useBalances } from "@/contexts/BalancesProvider"
+export { DomainProvider, useDomain } from "@/contexts/DomainProvider"
 export { TabsProvider, useTabs } from "@/contexts/TabsProvider"
 export { VaultProvider, useVaults } from "@/contexts/VaultProvider"
 export { WalletProvider } from "@/contexts/WalletProvider"
