+++ lib/hooks/useBlockchain.tsx
@@ -4,7 +4,7 @@ import { useCallback, useMemo } from "react"
 
 import { blockExplorerAddress, truncateAddress } from "@/lib/blockchain"
 
-export default function useBlockchain({ address }: { address: string }) {
+export const useBlockchain = ({ address }: { address: string }) => {
 	const displayAddress = useMemo(() => truncateAddress(address), [address])
 
 	const blockExplorer = useCallback(
@@ -16,3 +16,5 @@ export default function useBlockchain({ address }: { address: string }) {
 
 	return { address, displayAddress, blockExplorer }
 }
+
+export default useBlockchain
