+++ prisma/schema.prisma
@@ -1,4 +1,4 @@
-//🧩Don't Edit this file.✨Generated in Tue Mar 12 2024 22:12:06 GMT-0500 (Central Daylight Time)✨
+//🧩Don't Edit this file.✨Generated in Thu Mar 14 2024 02:35:03 GMT-0500 (Central Daylight Time)✨
 model Account {
   id                String  @id @default(cuid())
   userId            String
